App/requirements.txt
blinker==1.4
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.3
Flask==2.1.2
Flask-Mail==0.9.1
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
numpy==1.22.3
opencv-python==4.5.5.64
Pillow==9.1.0
python-dotenv==0.20.0
requests==2.27.1
SQLAlchemy==1.4.36
urllib3==1.26.9
Werkzeug==2.1.2
zipp==3.8.0
